backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.36
pydantic-settings==2.1.0
python-dotenv==1.0.0
google-generativeai==0.3.2
httpx==0.26.0
python-multipart==0.0.6
